add polling base for triggers
